backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
youtube-transcript-api
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
transformers
torch
yt-dlp
websockets
langchain_openai
langchain-text-splitters
langchain_core



